+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/AxisLabelTests.cs
@@ -6,10 +6,10 @@ internal class AxisLabelTests
     public void Test_AxisLabel_Rotation()
     {
         ScottPlot.Plot plt = new();
-        plt.BottomAxis.Label.Text = "Horizontal Axis";
-        plt.BottomAxis.Label.Rotation = 10;
-        plt.LeftAxis.Label.Text = "Vertical Axis";
-        plt.LeftAxis.Label.Rotation = -90 + 10;
+        plt.Axes.Bottom.Label.Text = "Horizontal Axis";
+        plt.Axes.Bottom.Label.Rotation = 10;
+        plt.Axes.Left.Label.Text = "Vertical Axis";
+        plt.Axes.Left.Label.Rotation = -90 + 10;
         plt.SaveTestImage();
     }
 }