+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/AxisTicksTests.cs
@@ -6,11 +6,11 @@ internal class AxisTicksTests
     public void Test_DateTimeTicks_Bottom()
     {
         Plot plt = new();
-        plt.BottomAxis.TickGenerator = new ScottPlot.TickGenerators.DateTimeAutomatic();
+        plt.Axes.Bottom.TickGenerator = new ScottPlot.TickGenerators.DateTimeAutomatic();
 
         DateTime dt1 = new(2023, 01, 01);
         DateTime dt2 = new(2024, 01, 01);
-        plt.SetAxisLimitsX(dt1.ToOADate(), dt2.ToOADate());
+        plt.Axes.SetLimitsX(dt1.ToOADate(), dt2.ToOADate());
 
         plt.SaveTestImage();
     }
@@ -20,11 +20,11 @@ internal class AxisTicksTests
     {
         Plot plt = new();
 
-        plt.LeftAxis.TickGenerator = new ScottPlot.TickGenerators.DateTimeAutomatic();
+        plt.Axes.Left.TickGenerator = new ScottPlot.TickGenerators.DateTimeAutomatic();
 
         DateTime dt1 = new(2023, 01, 01);
         DateTime dt2 = new(2024, 01, 01);
-        plt.SetAxisLimitsY(dt1.ToOADate(), dt2.ToOADate());
+        plt.Axes.SetLimitsY(dt1.ToOADate(), dt2.ToOADate());
 
         plt.SaveTestImage();
     }