+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/LayoutTests.cs
@@ -9,16 +9,16 @@ internal class LayoutTests
         plt.Add.Signal(ScottPlot.Generate.Sin());
         plt.Add.Signal(ScottPlot.Generate.Cos());
 
-        foreach (IAxis panel in plt.GetAxes())
+        foreach (IAxis panel in plt.Axes.GetAxes())
         {
             panel.ShowDebugInformation = true;
         }
 
-        plt.BottomAxis.Label.Text = "Horizontal Axis";
-        plt.LeftAxis.Label.Text = "Vertical Axis";
-        plt.RightAxis.Label.Text = "Secondary Vertical Axis";
-        plt.TopAxis.Label.Text = "Secondary Horizontal Axis";
-        plt.TitlePanel.Label.Text = "Title Panel";
+        plt.Axes.Bottom.Label.Text = "Horizontal Axis";
+        plt.Axes.Left.Label.Text = "Vertical Axis";
+        plt.Axes.Right.Label.Text = "Secondary Vertical Axis";
+        plt.Axes.Top.Label.Text = "Secondary Horizontal Axis";
+        plt.Axes.Title.Label.Text = "Title Panel";
 
         plt.SaveTestImage();
     }
@@ -30,17 +30,17 @@ internal class LayoutTests
         plt.Add.Signal(ScottPlot.Generate.Sin());
         plt.Add.Signal(ScottPlot.Generate.Cos());
 
-        foreach (IAxis panel in plt.GetAxes())
+        foreach (IAxis panel in plt.Axes.GetAxes())
         {
             panel.MinimumSize = 100;
             panel.ShowDebugInformation = true;
         }
 
-        plt.BottomAxis.Label.Text = "Horizontal Axis";
-        plt.LeftAxis.Label.Text = "Vertical Axis";
-        plt.RightAxis.Label.Text = "Secondary Vertical Axis";
-        plt.TopAxis.Label.Text = "Secondary Horizontal Axis";
-        plt.TitlePanel.Label.Text = "Title Panel";
+        plt.Axes.Bottom.Label.Text = "Horizontal Axis";
+        plt.Axes.Left.Label.Text = "Vertical Axis";
+        plt.Axes.Right.Label.Text = "Secondary Vertical Axis";
+        plt.Axes.Top.Label.Text = "Secondary Horizontal Axis";
+        plt.Axes.Title.Label.Text = "Title Panel";
 
         plt.SaveTestImage();
     }
@@ -52,17 +52,17 @@ internal class LayoutTests
         plt.Add.Signal(ScottPlot.Generate.Sin());
         plt.Add.Signal(ScottPlot.Generate.Cos());
 
-        foreach (IAxis panel in plt.GetAxes())
+        foreach (IAxis panel in plt.Axes.GetAxes())
         {
             panel.MaximumSize = 20;
             panel.ShowDebugInformation = true;
         }
 
-        plt.BottomAxis.Label.Text = "Horizontal Axis";
-        plt.LeftAxis.Label.Text = "Vertical Axis";
-        plt.RightAxis.Label.Text = "Secondary Vertical Axis";
-        plt.TopAxis.Label.Text = "Secondary Horizontal Axis";
-        plt.TitlePanel.Label.Text = "Title Panel";
+        plt.Axes.Bottom.Label.Text = "Horizontal Axis";
+        plt.Axes.Left.Label.Text = "Vertical Axis";
+        plt.Axes.Right.Label.Text = "Secondary Vertical Axis";
+        plt.Axes.Top.Label.Text = "Secondary Horizontal Axis";
+        plt.Axes.Title.Label.Text = "Title Panel";
 
         plt.SaveTestImage();
     }
@@ -74,18 +74,18 @@ internal class LayoutTests
         plt.Add.Signal(ScottPlot.Generate.Sin());
         plt.Add.Signal(ScottPlot.Generate.Cos());
 
-        foreach (IAxis panel in plt.GetAxes())
+        foreach (IAxis panel in plt.Axes.GetAxes())
         {
             panel.MinimumSize = 50;
             panel.MaximumSize = 50;
             panel.ShowDebugInformation = true;
         }
 
-        plt.BottomAxis.Label.Text = "Horizontal Axis";
-        plt.LeftAxis.Label.Text = "Vertical Axis";
-        plt.RightAxis.Label.Text = "Secondary Vertical Axis";
-        plt.TopAxis.Label.Text = "Secondary Horizontal Axis";
-        plt.TitlePanel.Label.Text = "Title Panel";
+        plt.Axes.Bottom.Label.Text = "Horizontal Axis";
+        plt.Axes.Left.Label.Text = "Vertical Axis";
+        plt.Axes.Right.Label.Text = "Secondary Vertical Axis";
+        plt.Axes.Top.Label.Text = "Secondary Horizontal Axis";
+        plt.Axes.Title.Label.Text = "Title Panel";
 
         plt.SaveTestImage();
     }
@@ -99,15 +99,15 @@ internal class LayoutTests
 
         // plottables use the standard X and Y axes by default
         var sig1 = myPlot.Add.Signal(Generate.Sin(51, mult: 0.01));
-        sig1.Axes.XAxis = myPlot.XAxis;
-        sig1.Axes.YAxis = myPlot.YAxis;
-        myPlot.YAxis.Label.Text = "Primary Y Axis";
+        sig1.Axes.XAxis = myPlot.Axes.Bottom;
+        sig1.Axes.YAxis = myPlot.Axes.Left;
+        myPlot.Axes.Left.Label.Text = "Primary Y Axis";
 
         // add a new plottable and tell it to use the right Y axis
         var sig2 = myPlot.Add.Signal(Generate.Cos(51, mult: 100));
-        sig2.Axes.XAxis = myPlot.XAxis;
-        sig2.Axes.YAxis = myPlot.RightAxis;
-        myPlot.RightAxis.Label.Text = "Secondary Y Axis";
+        sig2.Axes.XAxis = myPlot.Axes.Bottom;
+        sig2.Axes.YAxis = myPlot.Axes.Right;
+        myPlot.Axes.Right.Label.Text = "Secondary Y Axis";
 
         // Use fixed layout
         myPlot.Layout.Fixed(new PixelPadding(100));